+++ packages/frontend/core/src/components/affine/affine-error-boundary/error-basic/info-logger.tsx
@@ -1,7 +1,7 @@
 import {
   currentWorkspaceAtom,
   workspaceListAtom,
-} from '@affine/workspace/atom';
+} from '@affine/core/modules/workspace';
 import { useAtomValue } from 'jotai/react';
 import { useEffect } from 'react';
 import { useLocation, useParams } from 'react-router-dom';
