+++ packages/frontend/core/src/components/affine/create-workspace-modal/index.tsx
@@ -5,11 +5,11 @@ import {
   Modal,
 } from '@affine/component/ui/modal';
 import { useAsyncCallback } from '@affine/core/hooks/affine-async-hooks';
+import { workspaceManagerAtom } from '@affine/core/modules/workspace';
 import { DebugLogger } from '@affine/debug';
 import { apis } from '@affine/electron-api';
 import { WorkspaceFlavour } from '@affine/env/workspace';
 import { useAFFiNEI18N } from '@affine/i18n/hooks';
-import { workspaceManagerAtom } from '@affine/workspace/atom';
 import { getCurrentStore } from '@toeverything/infra/atom';
 import {
   buildShowcaseWorkspace,
