+++ packages/frontend/core/src/components/affine/share-page-modal/index.tsx
@@ -1,7 +1,7 @@
 import { useAsyncCallback } from '@affine/core/hooks/affine-async-hooks';
+import { workspaceManagerAtom } from '@affine/core/modules/workspace';
 import { WorkspaceFlavour } from '@affine/env/workspace';
 import type { Workspace } from '@affine/workspace';
-import { workspaceManagerAtom } from '@affine/workspace/atom';
 import type { Page } from '@blocksuite/store';
 import { useAtomValue } from 'jotai';
 import { useState } from 'react';
