+++ packages/frontend/core/src/components/blocksuite/block-suite-header-title/operation-menu.tsx
@@ -6,9 +6,9 @@ import {
   MenuSeparator,
 } from '@affine/component/ui/menu';
 import { useBlockSuitePageMeta } from '@affine/core/hooks/use-block-suite-page-meta';
+import { waitForCurrentWorkspaceAtom } from '@affine/core/modules/workspace';
 import { WorkspaceFlavour } from '@affine/env/workspace';
 import { useAFFiNEI18N } from '@affine/i18n/hooks';
-import { waitForCurrentWorkspaceAtom } from '@affine/workspace/atom';
 import { assertExists } from '@blocksuite/global/utils';
 import {
   DuplicateIcon,
