+++ packages/frontend/core/src/components/top-tip.tsx
@@ -1,11 +1,11 @@
 import { BrowserWarning } from '@affine/component/affine-banner';
 import { LocalDemoTips } from '@affine/component/affine-banner';
 import { useAsyncCallback } from '@affine/core/hooks/affine-async-hooks';
+import { workspaceManagerAtom } from '@affine/core/modules/workspace';
 import { WorkspaceFlavour } from '@affine/env/workspace';
 import { Trans } from '@affine/i18n';
 import { useAFFiNEI18N } from '@affine/i18n/hooks';
 import type { Workspace } from '@affine/workspace';
-import { workspaceManagerAtom } from '@affine/workspace/atom';
 import { useAtomValue, useSetAtom } from 'jotai';
 import { useCallback, useState } from 'react';
