+++ packages/frontend/core/src/hooks/current/use-current-page.ts
@@ -1,5 +1,5 @@
 import { useBlockSuiteWorkspacePage } from '@affine/core/hooks/use-block-suite-workspace-page';
-import { waitForCurrentWorkspaceAtom } from '@affine/workspace/atom';
+import { waitForCurrentWorkspaceAtom } from '@affine/core/modules/workspace';
 import { useAtomValue } from 'jotai';
 
 import { currentPageIdAtom } from '../../atoms/mode';
