+++ packages/frontend/core/src/modules/workspace/index.ts
@@ -1,0 +1,1 @@
+export * from './atoms';
