+++ packages/frontend/core/src/pages/workspace/detail-page/detail-page.tsx
@@ -6,9 +6,9 @@ import {
 import { ResizePanel } from '@affine/component/resize-panel';
 import { useBlockSuitePageMeta } from '@affine/core/hooks/use-block-suite-page-meta';
 import { useWorkspaceStatus } from '@affine/core/hooks/use-workspace-status';
+import { waitForCurrentWorkspaceAtom } from '@affine/core/modules/workspace';
 import { WorkspaceSubPath } from '@affine/env/workspace';
 import { globalBlockSuiteSchema, SyncEngineStep } from '@affine/workspace';
-import { waitForCurrentWorkspaceAtom } from '@affine/workspace/atom';
 import type { AffineEditorContainer } from '@blocksuite/presets';
 import type { Page, Workspace } from '@blocksuite/store';
 import { appSettingAtom } from '@toeverything/infra/atom';
