+++ tests/storybook/src/stories/image-preview-modal.stories.tsx
@@ -1,6 +1,6 @@
 import { BlockSuiteEditor } from '@affine/component/block-suite-editor';
 import { ImagePreviewModal } from '@affine/core/components/image-preview';
-import { waitForCurrentWorkspaceAtom } from '@affine/workspace/atom';
+import { waitForCurrentWorkspaceAtom } from '@affine/core/modules/workspace';
 import type { Page } from '@blocksuite/store';
 import type { Meta } from '@storybook/react';
 import { initEmptyPage } from '@toeverything/infra/blocksuite';
